requirements.txt
robotframework>=3.0
selenium>=3.141.0
requests>=2.25.1
